move timer into a submodule of `time`
this is because it would be nice to add stuff like timeout etc
