+++ maitake/src/lib.rs
@@ -17,4 +17,4 @@ pub mod future;
 pub mod scheduler;
 pub mod sync;
 pub mod task;
-pub mod timer;
+pub mod time;
