+++ maitake/src/time.rs
@@ -1,0 +1,6 @@
+//! Time utilities.
+mod sleep;
+mod timer;
+
+pub use self::sleep::Sleep;
+pub use self::timer::Timer;
